[refactor] #58 NicknameScreen -> LoginScreen으로 코드 이동 및 파일 삭제
